add sms reader

+++ main/DefaultCommands.h
@@ -23,7 +23,7 @@ enum class DEFAULT_COMMANDS
 
 struct DateTime
 {
-    uint16_t year{0};
+    uint8_t year{0};
     uint8_t month{0}; //0-11
     uint8_t day{0}; //0-30
     uint8_t hours{0}; //0-23
